Break up collectors into multiple files
