+++ runtime/appruntime/metrics/prometheus_exporter.go
@@ -13,7 +13,7 @@ func init() {
 			return cfg.Prometheus != nil
 		},
 		newExporter: func(m *Manager) exporter {
-			containerMetadata, err := metadata.GetContainerMetadata(m.cfg.Runtime, m.rootLogger)
+			containerMetadata, err := metadata.GetContainerMetadata(m.cfg.Runtime)
 			if err != nil {
 				m.rootLogger.Err(err).Msg("unable to initialize metrics exporter: error getting container metadata")
 				return nil
